* 'master' of github.com:DanielSeehausen/bear:
  Updated sessions, added methods to users controller, and added more routes
  created bear_api_development database. Migrated. Fixed table types. Incorporated Associations
  completed rake task seeding 120 days from s&p
  Fixed Game Rounds Controller
  Added more Auth info
  Authentication and Sessions Controller
  manipulated json to return S&P index for 100 days
  added yahoo finance api wrapper
  created new resources within server directory
  fixed file structure for rails api to be in server file after saving
  fixed file structure for rails api to be in server file

+++ server/bear_api/db/migrate/20170417203425_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[5.0]
       t.string :username
       t.string :password
       t.integer :games_played
-      t.double :cumulative_performance
+      t.float :cumulative_performance
 
       t.timestamps
     end
